custom cursor and seperated the save and load screens
