+++ saves/save1.txt
@@ -1,1 +1,1 @@
-{"username": "mlp", "score": 0}
+{"username": "mlp", "score": 3}
